+++ packages/core/src/payment/payment-method-action-creator.spec.ts
@@ -61,9 +61,7 @@ describe('PaymentMethodActionCreator', () => {
                 { type: PaymentMethodActionType.LoadPaymentMethodsRequested },
                 {
                     type: PaymentMethodActionType.LoadPaymentMethodsSucceeded,
-                    payload: paymentMethodsResponse.body.filter(
-                        (method) => method.id !== 'applepay',
-                    ),
+                    payload: paymentMethodsResponse.body,
                     meta: {
                         deviceSessionId: paymentMethodsResponse.headers['x-device-session-id'],
                         sessionHash: paymentMethodsResponse.headers['x-session-hash'],