+++ packages/core/src/payment/strategies/apple-pay/index.ts
@@ -1,5 +1,0 @@
-export {
-    ApplePayWindow,
-    assertApplePayWindow,
-    default as isApplePayWindow,
-} from './is-apple-pay-window';